Extracts background functions into own module
