+++ src/js/background_lists/default-backgrounds.js
@@ -4,7 +4,7 @@ function _importAll(r) {
     return r.keys().map(r);
 }
 
-const images = _importAll(require.context('../images/default_backgrounds/', false, /\.(png|jpe?g|svg)$/));
+const images = _importAll(require.context('../../images/default_backgrounds/', false, /\.(png|jpe?g|svg)$/));
 
 const default_backgrounds = images.map(function(img){
     return { 'url': img.default };
